website-scraper-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "website-scraper-mcp"
version = "0.1.0"
description = "MCP server for scraping websites to extract logos, UI styling, and color palettes"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.1",
    "playwright>=1.32.1",
    "beautifulsoup4>=4.11.2",
    "pillow>=9.4.0",
    "colorthief>=0.2.1",
    "aiohttp>=3.8.4",
    "pydantic>=1.10.7",
    "python-multipart>=0.0.6",
    "fastmcp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.2.0",
]

[project.scripts]
website-scraper-mcp = "website_scraper_mcp.main:main"

[tool.setuptools]
packages = ["website_scraper_mcp"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true